dryice-python/pyproject.toml
[build-system]
requires = ["maturin>=1.11,<2.0"]
build-backend = "maturin"

[project]
name = "dryice-python"
description = "Python bindings for the dryice high-throughput genomic record container"
requires-python = ">=3.9"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dynamic = ["version"]

[tool.maturin]
features = ["pyo3/abi3-py39"]

[dependency-groups]
dev = [
    "biopython>=1.85",
    "pytest>=8.4.2",
    "ruff>=0.15.6",
    "ty>=0.0.23",
]
